refactor(structure): group job modules into _jobs/ subpackage
Move _jobs_audit.py + _jobs_plugin.py into a new _jobs/ subpackage, dropping the flat root .py count from 15 to 13 (excluding __init__/__main__) so PS-108b (>15 threshold) keeps a safe margin as the package grows. Mirrors the two test files into tests/.../_jobs/, updates the scitex_dev.jobs entry point and every importer to the new dotted path. No external importers (only commented-out example lines in scitex-dev).
